+++ components/card/index.ts
@@ -1,0 +1,2 @@
+export { default as Card } from './Card';
+export { default as HikingCard } from './HikingCard';
